rovo/requirements.txt
playwright==1.41.1
python-dotenv==1.0.0
crewai==0.28.8
agentops==0.3.14
google-generativeai==0.3.2
asyncio-compat==0.1.2